requirements.txt
Flask
openai-whisper
torch
requests
markdown2
google-api-python-client
google-auth
google-auth-oauthlib
mistune
gunicorn